Move ManifestFormat to plugins module
