+++ test/service/test_manifest_async.py
@@ -35,6 +35,9 @@ from azul.logging import (
 from azul.modules import (
     load_app_module,
 )
+from azul.plugins import (
+    ManifestFormat,
+)
 from azul.service import (
     Filters,
 )
@@ -48,7 +51,6 @@ from azul.service.manifest_service import (
     CachedManifestNotFound,
     CachedManifestSourcesChanged,
     Manifest,
-    ManifestFormat,
     ManifestPartition,
     ManifestService,
 )
